backend-b/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
groq>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
